Conditionally copy grpc binaries
We only need to copy the grpc binaries manually until builds using the grpc library include them by default. That change will come in a future refactor of grpc. We're anticipating that change now so that users will not have to update their projects or build process later once the grpc library is refactored to reference binaries in a build-friendly way.