Conditionally copy grpc binaries (#78)
* Conditionally copy grpc binaries

We only need to copy the grpc binaries manually until builds using the grpc library include them by default. That change will come in a future refactor of grpc. We're anticipating that change now so that users will not have to update their projects or build process later once the grpc library is refactored to reference binaries in a build-friendly way.

* Checking the build files for bins instead of folder

Removing an assumption that binaries are copied to the default support folder (<destination>\data). Instead checking from Files Generated by Build

* Update Post-Build Action in the LDO example

Keeping the LDO example measurement up-to-date with the latest Post-Build Action code