Removing vi.lib dependencies from UIs
The latest GPlugin is now a packed library and cannot provide any of vi.lib to the Measurement UIs that it hosts. As such as need to make our template and examples not rely on vi.lib for them to be minimally runnable. The only place we used vi.lib components in our basically empty UIs is an empty case structure than handled and cleared errors in an empty case structure. We will follow up later with guidance on how to create more complicated Measurement UIs that can still run in the LV runtime environment.