Fix issue #102: incorrect scripting of `Get Measurement Details` into projects with existing measurements (#103)
* Fix bug that could lead measurement scripter to assign incorrect Measurement Details (display name, class) when generating into a project that already contained an existing measurement.

Problem was that the scripter was finding the first `Get Measurement Details` VI, which could collide with an earlier measurement... Fix is to ensure that it looks within the .lvlib for the newly-generated measurement.  Added utility VI for finding the library & refactored Create Build Spec VI to also leverage it.  Cleaned up some leaky reference handling along the way.